usb: dwc3: Fix OTG role switch support for AM62
Fix OTG role switch support for AM62.
The overlying subsystem of the USB should pass on the OTG role-switch
with an extcon for the switch. But it doesn't do that, and jumps out
of there with the return. The extcon setup comes afterwards. This is
then not called and the switch with the ID pin via extcon does not work.

Signed-off-by: Stephan Dünner <stephan.duenner@bytesatwork.ch>